requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
attrs==17.4.0
Automat==0.6.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
blinker==1.4
certifi==2018.10.15
chardet==3.0.4
click==6.7
colorama==0.3.7
configparser==3.5.0
constantly==15.1.0
cryptography==2.1.4
decorator==4.3.0
Django==1.11
django-cors-headers==2.4.0
djangorestframework==3.8.2
enum34==1.1.6
fasteners==0.12.0
futures==3.2.0
html5lib==0.999999999
http-ece==1.0.5
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
ipaddress==1.0.17
ipython==5.8.0
ipython-genutils==0.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
lockfile==0.12.2
logilab-common==1.4.1
lxml==4.2.1
mccabe==0.6.1
monotonic==1.0
ndg-httpsclient==0.4.4
numpy==1.13.3
oauthlib==2.0.6
olefile==0.45.1
onesignal-sdk==0.1.1
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
piston-mini-client==0.7.5
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py-vapid==1.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
PyJWT==1.5.3
pyliblzma==0.5.3
pylint==1.9.3
pyOpenSSL==17.5.0
pyserial==3.4
pysqlite==1.0.1
python-apt==1.6.3
python-debian==0.1.32
python-ldap==3.0.0
pytz==2018.5
pyxdg==0.25
requests==2.19.1
rpm==4.14.1
scandir==1.9.0
service-identity==16.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
software-center-aptd-plugins==0.0.0
traitlets==4.3.2
Twisted==17.9.0
ubuntuone-client-data==14.4
unity-lens-photos==1.0
urlgrabber==3.10.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5
wrapt==1.10.11
yum-metadata-parser==1.1.4
zope.interface==4.3.2
